+++ src/app/(content-layout)/[groupId]/tasklist/_tasklist/components/ModalContents/RemoveTaskModalPopUp.tsx
@@ -1,18 +1,58 @@
 'use client';
-import { ModalPortal, ModalProvider, ModalTrigger } from '@/components/common/modal';
-import { RemoveTaskModalContent } from './RemoveTaskModalContent';
+import {
+  ModalContainer,
+  ModalDescription,
+  ModalFooter,
+  ModalHeading,
+  ModalOverlay,
+  ModalPortal,
+} from '@/components/common/modal';
+import Image from 'next/image';
+import Button from '@/components/common/Button';
+import useModalContext from '@/components/common/modal/core/useModalContext';
 
 interface Props {
   taskName: string;
+  modalId: string;
 }
 
-export default function RemoveTaskModal({ taskName }: Props) {
+export default function RemoveTaskModal({ taskName, modalId }: Props) {
+  const { closeModal } = useModalContext();
+
+  const handleClickDeleteTask = () => {
+    //삭제 요청
+  };
+
   return (
-    <ModalProvider>
-      <ModalTrigger modalId="1">삭제하기</ModalTrigger>
-      <ModalPortal modalId="1">
-        <RemoveTaskModalContent taskName={taskName} />
+    <>
+      <ModalPortal modalId={modalId}>
+        <ModalOverlay modalId={modalId} onClick={() => closeModal(`${modalId}`)}>
+          <ModalContainer>
+            <Image src="/icons/danger.icon.svg" alt="!" width={20} height={20} />
+            <ModalHeading className="mt-4 mb-2">
+              <p>{`'${taskName}'${(<br />)}할 일을 정말 삭제하시겠어요?`}</p>
+            </ModalHeading>
+            <ModalDescription className="text-md-md text-gray500 mb-6 w-full">
+              삭제 후에는 되돌릴 수 없습니다.
+            </ModalDescription>
+            <ModalFooter className="w-full">
+              <div className="flex w-full gap-2">
+                <Button
+                  variant="outline-gray"
+                  onClick={() => closeModal(`${modalId}`)}
+                  fontSize="16"
+                  size="fullWidth"
+                >
+                  닫기
+                </Button>
+                <Button variant="danger" onClick={handleClickDeleteTask} size="fullWidth">
+                  만들기
+                </Button>
+              </div>
+            </ModalFooter>
+          </ModalContainer>
+        </ModalOverlay>
       </ModalPortal>
-    </ModalProvider>
+    </>
   );
 }
